Add tests for 'getLiveUploadId' and 'listUploadIds' in server
